+++ packages/core/src/react/hooks/index.ts
@@ -2,6 +2,7 @@ export * from './useFetch';
 export * from './useFetchPost';
 export * from './useFetchPosts';
 export * from './useFetchSearch';
+export * from './useFetchSearchNative';
 export * from './types';
 export * from './useFetchTerms';
 export * from './useFetchAuthorArchive';