+++ wp/headless-wp/includes/classes/BaseToken.php
@@ -64,8 +64,12 @@ abstract class BaseToken {
 
 	/**
 	 * Decode capability tokens if present.
+	 *
+	 * @param string $token if set will use this instead of reading from the headers
+	 *
+	 * @return null|object
 	 */
-	protected static function get_payload_from_token() {
+	public static function get_payload_from_token( $token = '' ) {
 		// Get HTTP Authorization Header.
 		$header = isset( $_SERVER['HTTP_AUTHORIZATION'] )
 		? sanitize_text_field( wp_unslash( $_SERVER['HTTP_AUTHORIZATION'] ) )
@@ -76,18 +80,25 @@ abstract class BaseToken {
 			$header = sanitize_text_field( wp_unslash( $_SERVER['REDIRECT_HTTP_AUTHORIZATION'] ) );
 		}
 
+		if ( ! $header && isset( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] ) ) {
+			$header = sanitize_text_field( wp_unslash( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] ) );
+		}
+
 		// No Authorization Header is present.
-		if ( ! $header ) {
+		if ( ! $header && ! $token ) {
 			return null;
 		}
 
 		// Get and parse the token.
 		try {
-			list( $token ) = sscanf( $header, 'Bearer %s' );
+			if ( ! $token ) {
+				list( $token ) = sscanf( $header, 'Bearer %s' );
+			}
 
 			if ( empty( $token ) ) {
 				return null;
 			}
+
 			$payload = JWT::decode(
 				$token,
 				self::get_private_key(),