+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @headstartwp/core
 
+## 1.3.1
+
+### Patch Changes
+
+- 0bd8e415: Add ability to preview using an alternative authorization header
+
+## 1.3.0
+
+### Minor Changes
+
+- 8452279a: Implement WordPress native search endpoint
+
 ## 1.2.0
 
 ### Minor Changes